+++ Runtime/Cesium3DTileset.cs
@@ -68,9 +68,11 @@ namespace CesiumForUnity
             }
         }
 
+        internal static event Action OnSetShowCreditsOnScreen;
+
         [SerializeField]
         private bool _showCreditsOnScreen = false;
-        
+
         /// <summary>
         /// Whether or not to force this tileset's credits to be shown on the main screen. If false, the
         /// credits are usually only shown on a "Data Attribution" popup.
@@ -81,7 +83,11 @@ namespace CesiumForUnity
             set
             {
                 this._showCreditsOnScreen = value;
-                this.RecreateTileset();
+                this.SetShowCreditsOnScreen(this._showCreditsOnScreen);
+                if (Cesium3DTileset.OnSetShowCreditsOnScreen != null)
+                {
+                    Cesium3DTileset.OnSetShowCreditsOnScreen();
+                }
             }
         }
 
@@ -517,6 +523,19 @@ namespace CesiumForUnity
         // }
 
         [SerializeField]
+        private CesiumPointCloudShading _pointCloudShading;
+
+        /// <summary>
+        /// The CesiumPointCloudShading attached to this tileset. If the tileset
+        /// contains points, their appearance can be configured with the point
+        /// cloud shading parameters.
+        /// </summary>
+        public CesiumPointCloudShading pointCloudShading
+        {
+            get => this._pointCloudShading;
+        }
+
+        [SerializeField]
         private bool _suspendUpdate = false;
 
         /// <summary>
@@ -613,13 +632,15 @@ namespace CesiumForUnity
             }
         }
 
+        private partial void SetShowCreditsOnScreen(bool value);
+
         private partial void Start();
         private partial void Update();
         private partial void OnValidate();
 
         private partial void OnEnable();
         private partial void OnDisable();
-        
+
         /// <summary>
         /// Destroy and recreate the tilset. All tiles are unloaded, and then the tileset is reloaded
         /// based on the current view.