+++ native~/Runtime/src/CesiumBingMapsRasterOverlayImpl.cpp
@@ -11,6 +11,7 @@
 #include <DotNet/CesiumForUnity/Cesium3DTileset.h>
 #include <DotNet/CesiumForUnity/CesiumBingMapsRasterOverlay.h>
 #include <DotNet/CesiumForUnity/CesiumRasterOverlay.h>
+#include <DotNet/System/String.h>
 
 using namespace Cesium3DTilesSelection;
 using namespace DotNet;