+++ native~/Runtime/src/UnityTilesetExternals.cpp
@@ -9,14 +9,11 @@
 #include <CesiumAsync/SqliteCache.h>
 
 #include <DotNet/CesiumForUnity/CesiumCreditSystem.h>
+#include <DotNet/CesiumForUnity/CesiumRuntimeSettings.h>
 #include <DotNet/System/String.h>
 #include <DotNet/UnityEngine/Application.h>
 
 #include <memory>
-
-#if UNITY_EDITOR
-#include <DotNet/UnityEditor/EditorApplication.h>
-#endif
 
 using namespace Cesium3DTilesSelection;
 using namespace CesiumAsync;
@@ -29,12 +26,6 @@ namespace {
 std::shared_ptr<CachingAssetAccessor> pAccessor = nullptr;
 std::shared_ptr<UnityTaskProcessor> pTaskProcessor = nullptr;
 std::shared_ptr<CreditSystem> pCreditSystem = nullptr;
-#if UNITY_EDITOR
-// If a tileset is loading in the editor, it won't instantiate the
-// credit system prefab. A CreditSystem will be manually constructed
-// so that the TilesetExternals won't contain a nullptr.
-std::shared_ptr<CreditSystem> pEditorCreditSystem = nullptr;
-#endif
 
 const std::shared_ptr<CachingAssetAccessor>& getAssetAccessor() {
   if (!pAccessor) {
@@ -42,10 +33,18 @@ const std::shared_ptr<CachingAssetAccessor>& getAssetAccessor() {
         UnityEngine::Application::temporaryCachePath().ToStlString();
     std::string cacheDBPath = tempPath + "/cesium-request-cache.sqlite";
 
+    int32_t requestsPerCachePrune =
+        CesiumForUnity::CesiumRuntimeSettings::requestsPerCachePrune();
+    uint64_t maxItems = CesiumForUnity::CesiumRuntimeSettings::maxItems();
+
     pAccessor = std::make_shared<CachingAssetAccessor>(
         spdlog::default_logger(),
         std::make_shared<UnityAssetAccessor>(),
-        std::make_shared<SqliteCache>(spdlog::default_logger(), cacheDBPath));
+        std::make_shared<SqliteCache>(
+            spdlog::default_logger(),
+            cacheDBPath,
+            maxItems),
+        requestsPerCachePrune);
   }
   return pAccessor;
 }
@@ -59,16 +58,6 @@ const std::shared_ptr<UnityTaskProcessor>& getTaskProcessor() {
 
 const std::shared_ptr<CreditSystem>&
 getCreditSystem(const CesiumForUnity::Cesium3DTileset& tileset) {
-#if UNITY_EDITOR
-  if (UnityEngine::Application::isEditor() &&
-      !UnityEditor::EditorApplication::isPlaying()) {
-    if (!pEditorCreditSystem) {
-      pEditorCreditSystem = std::make_shared<CreditSystem>();
-    }
-    return pEditorCreditSystem;
-  }
-#endif
-
   // Get the credit system associated with the tileset.
   Cesium3DTilesetImpl& tilesetImpl = tileset.NativeImplementation();
   CesiumForUnity::CesiumCreditSystem creditSystem =
@@ -77,7 +66,7 @@ getCreditSystem(const CesiumForUnity::Cesium3DTileset& tileset) {
   // If the tileset does not already reference a credit system,
   // get the default one.
   if (creditSystem == nullptr) {
-    creditSystem = CesiumCreditSystemImpl::getDefaultCreditSystem();
+    creditSystem = CesiumForUnity::CesiumCreditSystem::GetDefaultCreditSystem();
     tilesetImpl.setCreditSystem(creditSystem);
   }